open beta

+++ RapChieuPhim/RapChieuPhim/Areas/Admin/Controllers/TaiKhoanController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -21,15 +22,36 @@ namespace RapChieuPhim.Areas.Admin.Controllers
         }
 
         // GET: Admin/TaiKhoan
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(String ten)
         {
-            var dPContext = _context.TaiKhoanModel.Include(t => t.NguoiDungModel);
-            return View(await dPContext.ToListAsync());
+            if (HttpContext.Session.GetString("tk") == null)
+            {
+                return RedirectToRoute(new { action = "Login", controller = "Home", area = "Admin" });
+            }
+            var dstk = from taikhoan in _context.TaiKhoanModel
+                       select taikhoan;
+            if (!String.IsNullOrWhiteSpace(ten))
+            {
+                dstk = dstk.Where(s => s.Ten_dang_nhap.Contains(ten));
+            }
+            //var dstk = (from s in _context.TaiKhoanModel
+            //            join c in _context.NguoiDungModel on s.NguoiDung_ID equals c.ID
+            //            select s).ToList();
+
+            ViewBag.dstk = dstk;
+            return View();
+
+            // var dPContext = _context.TaiKhoanModel.Include(t => t.NguoiDungModel);
+            //return View(await dPContext.ToListAsync());
         }
 
         // GET: Admin/TaiKhoan/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (HttpContext.Session.GetString("tk") == null)
+            {
+                return RedirectToRoute(new { action = "Login", controller = "Home", area = "Admin" });
+            }
             if (id == null)
             {
                 return NotFound();
@@ -50,9 +72,16 @@ namespace RapChieuPhim.Areas.Admin.Controllers
         [HttpGet]
         public IActionResult Create()
         {
-            ViewData["NguoiDung_ID"] = new SelectList(_context.NguoiDungModel, "ID", "ID");
+            if (HttpContext.Session.GetString("tk") == null)
+            {
+                return RedirectToRoute(new { action = "Login", controller = "Home", area = "Admin" });
+            }
             TaiKhoanModel taiKhoan = new TaiKhoanModel();
-            return PartialView("_TaiKhoanModelPartial", taiKhoan);
+
+            ViewBag.lstTaiKhoan = from l in _context.NguoiDungModel                            
+                                  select l;
+            ViewData["NguoiDung_ID"] = new SelectList(_context.NguoiDungModel, "ID", "ID");
+            return View();
         }
 
 
@@ -63,6 +92,10 @@ namespace RapChieuPhim.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ID,Ten_dang_nhap,Mat_khau,Loai_tai_khoan,NguoiDung_ID,Da_xoa")] TaiKhoanModel taiKhoanModel)
         {
+            if (HttpContext.Session.GetString("tk") == null)
+            {
+                return RedirectToRoute(new { action = "Login", controller = "Home", area = "Admin" });
+            }
             if (ModelState.IsValid)
             {
                 _context.Add(taiKhoanModel);
@@ -76,6 +109,10 @@ namespace RapChieuPhim.Areas.Admin.Controllers
         // GET: Admin/TaiKhoan/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (HttpContext.Session.GetString("tk") == null)
+            {
+                return RedirectToRoute(new { action = "Login", controller = "Home", area = "Admin" });
+            }
             if (id == null)
             {
                 return NotFound();
@@ -86,6 +123,9 @@ namespace RapChieuPhim.Areas.Admin.Controllers
             {
                 return NotFound();
             }
+            ViewBag.lstTaiKhoan = from l in _context.NguoiDungModel
+                                  select l;
+          
             ViewData["NguoiDung_ID"] = new SelectList(_context.NguoiDungModel, "ID", "ID", taiKhoanModel.NguoiDung_ID);
             return View(taiKhoanModel);
         }
@@ -97,6 +137,10 @@ namespace RapChieuPhim.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("ID,Ten_dang_nhap,Mat_khau,Loai_tai_khoan,NguoiDung_ID,Da_xoa")] TaiKhoanModel taiKhoanModel)
         {
+            if (HttpContext.Session.GetString("tk") == null)
+            {
+                return RedirectToRoute(new { action = "Login", controller = "Home", area = "Admin" });
+            }
             if (id != taiKhoanModel.ID)
             {
                 return NotFound();
@@ -129,6 +173,10 @@ namespace RapChieuPhim.Areas.Admin.Controllers
         // GET: Admin/TaiKhoan/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (HttpContext.Session.GetString("tk") == null)
+            {
+                return RedirectToRoute(new { action = "Login", controller = "Home", area = "Admin" });
+            }
             if (id == null)
             {
                 return NotFound();
@@ -150,6 +198,10 @@ namespace RapChieuPhim.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (HttpContext.Session.GetString("tk") == null)
+            {
+                return RedirectToRoute(new { action = "Login", controller = "Home", area = "Admin" });
+            }
             var taiKhoanModel = await _context.TaiKhoanModel.FindAsync(id);
             _context.TaiKhoanModel.Remove(taiKhoanModel);
             await _context.SaveChangesAsync();
